refactor: credentials

+++ src/credentials/credentials.repository.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { CredentialDto } from './dto/credential.dto';
 import { PrismaService } from '@/prisma/prisma.service';
 import { Credential } from '@prisma/client';
-import Cryptr from 'cryptr'; 
+const Cryptr = require('cryptr');
 
 @Injectable()
 export class CredentialsRepository {
@@ -22,41 +22,25 @@ export class CredentialsRepository {
   }
 
   findCredentialByUserIdAndTitle(userId: number, title: string) {
-    return this.prisma.credential.findUnique({
-      where: {
-        userId_title: {
-          userId,
-          title,
-        },
-      },
-    });
+    return this.prisma.credential.findUnique({ where: { userId_title: { userId, title } } });
   }
 
   async findAllCredentialsByUserId(userId: number) {
-    const credentials = await this.prisma.credential.findMany({
-      where: { userId },
-    });
-
+    const credentials = await this.prisma.credential.findMany({ where: { userId } });
     return this.decryptCredentialsPassword(credentials);
   }
 
   async findCredentialById(id: number) {
-    const credential = await this.prisma.credential.findFirst({
-      where: { id },
-    });
+    const credential = await this.prisma.credential.findFirst({ where: { id } });
     return credential ? this.decryptCredentialsPassword([credential]) : [];
   }
 
   deleteCredential(id: number) {
-    return this.prisma.credential.delete({
-      where: { id },
-    });
+    return this.prisma.credential.delete({ where: { id } });
   }
 
   deleteAllCredentials(userId: number) {
-    return this.prisma.credential.deleteMany({
-      where: { userId },
-    });
+    return this.prisma.credential.deleteMany({ where: { userId } });
   }
 
   private decryptCredentialsPassword(credentials: Credential[]) {
